requirements.txt
Django==2.2.11
Pillow==7.1.0
psycopg2==2.8.4
django-allauth==0.41.0
django-braces==1.14.0
django-crispy-forms==1.9.0
django-imagekit==4.0.1
django-heroku==0.3.1
gunicorn==20.0.4
